FCN-244 organize files better
