+++ packages/nxtcm-rosa-hcp-wizard/src/ROSAHCPWizardBody.tsx
@@ -12,7 +12,7 @@ import { Review } from './Steps/Review/Review';
 import type { YamlEditorHandle } from './Steps/YamlEditor/RosaHcpYamlEditorStep';
 import { createRosaHcpWizardFooter } from './Footer/RosaHcpWizardFooter';
 import { RosaHcpYamlEditorFooter } from './Footer/RosaHcpYamlEditorFooter';
-import { useWizardFieldMetaChangeEffects } from './hooks/useWizardFieldMetaChangeEffects';
+import { useWizardFieldMetaChangeEffects } from './fieldMetaChangeEffects/useWizardFieldMetaChangeEffects';
 import { useRosaHcpWizardStrings } from './stringsProvider/RosaHcpWizardStringsContext';
 import { STEP_IDS } from './constants';
 import type { RosaHCPWizardProps } from './types';
