+++ packages/nxtcm-rosa-hcp-wizard/src/Steps/BasicSetup/RolesAndPolicies/RolesAndPolicies.spec-helpers.tsx
@@ -15,13 +15,13 @@ import {
   type ROSAHCPWizardData,
   type RolesResource,
 } from '../../../types';
-import { WizardFieldMetaChangeEffects } from '../../../hooks/WizardFieldMetaChangeEffects';
 import { RolesAndPolicies } from './RolesAndPolicies';
 import fixtures from '../../../ROSAHCPWizard.fixtures';
 import { clusterValidationSchema } from '../../../yupSchemas';
 import type { ValidationSchemaContext } from '../../../yupSchemas/types';
 import { defaultRosaHcpWizardValidatorStrings } from '../../../stringsProvider/rosaHcpWizardStrings.defaults';
 import { withRosaCt } from '../../../components/WizFields/wizFieldCtSpecHelpers';
+import { WizardFieldMetaChangeEffectsCtHarness } from '../../../rosaHcpWizardCtSpecHelpers';
 
 /** Defaults aligned with {@link ROSAHCPWizardBody} so the composed Yup schema resolves consistently in CT. */
 const DEFAULT_ROSA_HCP_CT_FORM_VALUES: Partial<ROSAHCPCluster> = {
@@ -118,7 +118,7 @@ export const RolesAndPoliciesMount: React.FC<RolesAndPoliciesMountProps> = ({
   return withRosaCt(
     <FormProvider {...methods}>
       <Form>
-        <WizardFieldMetaChangeEffects wizardData={wizardData} />
+        <WizardFieldMetaChangeEffectsCtHarness wizardData={wizardData} />
         <RolesAndPolicies roles={rolesProps} oidcConfig={oidcProps} />
       </Form>
     </FormProvider>
