+++ packages/nxtcm-rosa-hcp-wizard/src/fieldMetaChangeEffects/applyWizardFieldMetaChangeEffects.test.ts
@@ -4,9 +4,9 @@ import { applyWizardFieldMetaChangeEffects } from './applyWizardFieldMetaChangeE
 import { resetFieldsToDefaultValues } from './resetFieldsToDefaultValues';
 import { syncFieldsOnSourceChange } from './syncFieldsOnSourceChange';
 import { applyWizardFieldDerivedSync } from './wizardFieldDerivedSyncs';
-import type { ROSAHCPCluster, ROSAHCPWizardData } from './types';
-import fixtures from './ROSAHCPWizard.fixtures';
-import { getWizardFieldSyncsForSourceField } from './yupSchemas';
+import type { ROSAHCPCluster, ROSAHCPWizardData } from '../types';
+import fixtures from '../ROSAHCPWizard.fixtures';
+import { getWizardFieldSyncsForSourceField } from '../yupSchemas';
 
 const autoscalingSyncRules = getWizardFieldSyncsForSourceField('autoscaling');
