+++ packages/nxtcm-rosa-hcp-wizard/src/fieldMetaChangeEffects/applyWizardFieldMetaChangeEffects.ts
@@ -4,14 +4,14 @@ import { hasDerivedSyncSourceValue, hasRefetchableStringValue } from './wizardFi
 import { resetFieldsToDefaultValues } from './resetFieldsToDefaultValues';
 import { syncFieldsOnSourceChange } from './syncFieldsOnSourceChange';
 import { applyWizardFieldDerivedSync } from './wizardFieldDerivedSyncs';
-import type { ROSAHCPCluster, ROSAHCPWizardData } from './types';
+import type { ROSAHCPCluster, ROSAHCPWizardData } from '../types';
 import {
   getWizardFieldDerivedSyncKeyForSourceField,
   getWizardFieldResetsForSourceField,
   getWizardFieldSyncsForSourceField,
   getWizardResourceRefetchesForSourceField,
-} from './yupSchemas';
-import type { WizardFormFieldName, WizardResourceRefetchOnChange } from './yupSchemas/types';
+} from '../yupSchemas';
+import type { WizardFormFieldName, WizardResourceRefetchOnChange } from '../yupSchemas/types';
 
 export type ApplyWizardFieldMetaChangeEffectsArgs = {
   sourceField: WizardFormFieldName;
