+++ packages/nxtcm-rosa-hcp-wizard/src/fieldMetaChangeEffects/resetFieldsToDefaultValues.test.ts
@@ -1,6 +1,7 @@
 import type { UseFormSetValue } from 'react-hook-form';
-import type { ROSAHCPCluster } from './types';
-import { getClusterValidationSchemaDefaultValues } from './yupSchemas';
+
+import type { ROSAHCPCluster } from '../types';
+import { getClusterValidationSchemaDefaultValues } from '../yupSchemas';
 import { resetFieldsToDefaultValues } from './resetFieldsToDefaultValues';
 
 describe('resetFieldsToDefaultValues', () => {
