+++ packages/nxtcm-rosa-hcp-wizard/src/fieldMetaChangeEffects/wizardFieldDerivedSyncs.ts
@@ -1,9 +1,9 @@
 import type { UseFormSetValue } from 'react-hook-form';
 
-import { resolveSelectedVpc } from './helpers';
-import type { ROSAHCPCluster, ROSAHCPWizardData, Role, VPC } from './types';
-import type { WizardFieldDerivedSyncKey } from './yupSchemas/types';
-import type { WizardFieldDerivedSyncEntry } from './yupSchemas/wizardFieldMetaChangeRegistry';
+import { resolveSelectedVpc } from '../helpers';
+import type { ROSAHCPCluster, ROSAHCPWizardData, Role, VPC } from '../types';
+import type { WizardFieldDerivedSyncKey } from '../yupSchemas/types';
+import type { WizardFieldDerivedSyncEntry } from '../yupSchemas/wizardFieldMetaChangeRegistry';
 
 /** Non-empty string suitable for resource refetch args and string-backed derived sync sources. */
 export function hasRefetchableStringValue(value: unknown): value is string {
